Move tests from extension
